\name{bodyfat}
\docType{data}
\alias{bodyfat}
\title{Bodyfat data}
\description{A data frame containing the estimates of the percentage of body fat determined by underwater weighing and various body circumference measurements for 252 men}
\usage{data(bodyfat)}
\format{Percent body fat using Brozek's equation: 457/Density - 414.2}
\source{http://lib.stat.cmu.edu/datasets/bodyfat}

