% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{summary.mmfit}
\alias{summary.mmfit}
\title{summary.mmfit function and plot.mmfit function}
\usage{
\method{summary}{mmfit}(x, g, start, type)
}
\arguments{
\item{x}{a set of sample data}

\item{g}{moment function}

\item{start}{initial values to start with}

\item{type}{distribution type. It take a value from "Poisson", "power law", "Gamma", "Beta", mixture of "2Poissons", and "2Exponential".}
}
\description{
We can use summary.mmfit(x, g, start, type) to obtain estimated distribution parameters.
We use plot.mmfit(x, g, start, type) to plot fitting curve and empirical CDF with K-S confidence band.
}

